# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "alphauniverse",
    "version": "0.1.0"
  },
  "paths": {
    "/meta": {
      "get": {
        "summary": "Get Meta",
        "operationId": "get_meta",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              }
            }
          }
        }
      }
    },
    "/projections/{projection}": {
      "get": {
        "summary": "Get Projections",
        "operationId": "get_projections",
        "parameters": [
          {
            "name": "projection",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "mean",
                "full"
              ],
              "type": "string",
              "title": "Projection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/downloads/{role}": {
      "get": {
        "summary": "Download Artifact",
        "operationId": "download_artifact",
        "parameters": [
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "encoded",
                "codebook",
                "tokens"
              ],
              "type": "string",
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/galaxy/{galaxy}": {
      "get": {
        "summary": "Get Galaxy",
        "operationId": "get_galaxy",
        "parameters": [
          {
            "name": "galaxy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Galaxy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Galaxy"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/galaxy/{galaxy}/image": {
      "get": {
        "summary": "Get Image",
        "operationId": "get_image",
        "parameters": [
          {
            "name": "galaxy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Galaxy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/galaxy/{galaxy}/image/tokens": {
      "get": {
        "summary": "Get Image Tokens",
        "operationId": "get_image_tokens",
        "parameters": [
          {
            "name": "galaxy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Galaxy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/galaxy/{galaxy}/spectrum": {
      "get": {
        "summary": "Get Spectrum",
        "operationId": "get_spectrum",
        "parameters": [
          {
            "name": "galaxy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Galaxy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/vnd.apache.arrow.stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/galaxy/{galaxy}/spectrum/tokens": {
      "get": {
        "summary": "Get Spectrum Tokens",
        "operationId": "get_spectrum_tokens",
        "parameters": [
          {
            "name": "galaxy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Galaxy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Get Search",
        "operationId": "get_search",
        "parameters": [
          {
            "name": "galaxy",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Galaxy"
            }
          },
          {
            "name": "p",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "minimum": 0,
                "exclusiveMaximum": 576
              },
              "maxItems": 576,
              "default": [],
              "title": "P"
            }
          },
          {
            "name": "s",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "minimum": 0,
                "exclusiveMaximum": 272
              },
              "maxItems": 272,
              "default": [],
              "title": "S"
            }
          },
          {
            "name": "matches",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 128,
              "minimum": 1,
              "default": 32,
              "title": "Matches"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/vnd.apache.arrow.stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search/text": {
      "get": {
        "summary": "Get Text Search",
        "operationId": "get_text_search",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 500,
              "title": "Text"
            }
          },
          {
            "name": "matches",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 128,
              "minimum": 1,
              "default": 32,
              "title": "Matches"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextMatches"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Galaxy": {
        "properties": {
          "ls": {
            "type": "boolean",
            "title": "Ls"
          },
          "hsc": {
            "type": "boolean",
            "title": "Hsc"
          },
          "desi": {
            "type": "boolean",
            "title": "Desi"
          },
          "sdss": {
            "type": "boolean",
            "title": "Sdss"
          },
          "gz10": {
            "type": "boolean",
            "title": "Gz10"
          },
          "provabgs": {
            "type": "boolean",
            "title": "Provabgs"
          }
        },
        "type": "object",
        "required": [
          "ls",
          "hsc",
          "desi",
          "sdss",
          "gz10",
          "provabgs"
        ],
        "title": "Galaxy"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Meta": {
        "properties": {
          "author": {
            "type": "string",
            "title": "Author"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "galaxies": {
            "type": "integer",
            "title": "Galaxies"
          },
          "grid": {
            "type": "integer",
            "title": "Grid"
          },
          "spectrum_origin": {
            "type": "number",
            "title": "Spectrum Origin"
          },
          "spectrum_width": {
            "type": "number",
            "title": "Spectrum Width"
          },
          "morphologies": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Morphologies"
          }
        },
        "type": "object",
        "required": [
          "author",
          "id",
          "galaxies",
          "grid",
          "spectrum_origin",
          "spectrum_width",
          "morphologies"
        ],
        "title": "Meta"
      },
      "TextMatches": {
        "properties": {
          "galaxies": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Galaxies"
          },
          "scores": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Scores"
          }
        },
        "type": "object",
        "required": [
          "galaxies",
          "scores"
        ],
        "title": "TextMatches"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}